requirements.txt
fastapi
uvicorn
sqlalchemy
aiohttp
discord.py
